backend/api/serializers.py
from rest_framework import serializers
from api.models import Todo


class TodoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        fields = ('id', 'title', 'description', 'completed', 'date_added', 'date_completed')