requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
autopep8==1.6.0
    # via -r requirements.in
django==4.0.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.13.0
    # via -r requirements.in
django-dotenv==1.4.2
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pycodestyle==2.8.0
    # via autopep8
pytz==2022.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via autopep8
tzdata==2022.1
    # via django
